+++ packages/api/src/settings/transformers.ts
@@ -1,7 +1,9 @@
 import { SettingDocument, SettingDto } from './types';
 
 const settingDocumentToDto = (setting: SettingDocument): SettingDto => {
-  const { _id, key, value, valueRealized, type, label, description } = setting;
+  const { _id, key, value, valueRealized, type, label, description, group } =
+    setting;
+
   return {
     _id,
     key,
@@ -10,6 +12,7 @@ const settingDocumentToDto = (setting: SettingDocument): SettingDto => {
     type,
     label,
     description,
+    group,
   };
 };